[CODER] Implemented DynamicLossScaler for AMP training
- Created DynamicLossScaler class with automatic scale adjustment
- Implemented growth/backoff logic based on overflow detection
- Added comprehensive statistics tracking (DynamicScalerStats)
- Updated GradScaler to use new DynamicLossScaler API
- Fixed Half.cs casting errors
- Created comprehensive unit tests for all features
- Supports min/max scale bounds and growth interval configuration

+++ src/MLFramework/Amp/Half.cs
@@ -136,9 +136,9 @@ namespace MLFramework.Amp
         {
             unsafe
             {
-                uint sign = (uint)(half >> 15) << 31;
+                uint sign = ((uint)half >> 15) << 31;
                 uint exponent = (uint)((half >> 10) & 0x1F);
-                uint mantissa = half & 0x3FF;
+                uint mantissa = (uint)(half & 0x3FF);
 
                 if (exponent == 0)
                 {
